refactor: network model 파일 위치 변경에 따른  (#19)
